Truth Table Generator
A GUI program. It takes a logic expression and draws the
corresponding truth table. Then user can find the truth value of the
expression given the values of its
propositions, determine if the expression is a tautology or a
contradiction, and test if two
expressions are equivalent by comparing the truth table for both of
them.